refact: clean server.rs from log/ tls config
